to solve the following errors logged in Catalina.log:
—org.apache.catalina.loader.WebappClassLoader.clearReferencesJdbc The
web application [] registered the JDBC driver [com.mysql.jdbc.Driver]
but failed to unregister it when the web application was stopped. To
prevent a memory leak, the JDBC Driver has been forcibly unregistered.

—The web application [] appears to have started a thread named [Health
Check Thread for
LDAPConnectionPool(serverSet=SingleServerSet(server=10.4.6.12:1389),
maxConnections=100)] but has failed to stop it. This is very likely to
create a memory leak.

—org.apache.catalina.loader.WebappClassLoader.checkThreadLocalMapForLeak
s The web application [] created a ThreadLocal with key of type
[java.lang.ThreadLocal] (value [java.lang.ThreadLocal@d5a90a0]) and a
value of type [com.unboundid.asn1.ASN1Buffer] (value
[com.unboundid.asn1.ASN1Buffer@16385713]) but failed to remove it when
the web application was stopped. Threads are going to be renewed over
time to try and avoid a probable memory leak.